restructure files, adds a few comments/docs
