--- src/document.jl
@@ -1,4 +1,0 @@
-type Document
-    data::Vector{UInt8}
-    blocks::Vector{Any}
-end
