+++ src/languageserverinstance.jl
@@ -1,3 +1,55 @@
+const TextDocumentSyncKind = Dict("None"=>0, "Full"=>1, "Incremental"=>2)
+
+type CompletionOptions 
+    resolveProvider::Bool
+    triggerCharacters::Vector{String}
+end
+
+type SignatureHelpOptions
+    triggerCharacters::Vector{String}
+end
+
+type ServerCapabilities
+    textDocumentSync::Int
+    hoverProvider::Bool
+    completionProvider::CompletionOptions
+    definitionProvider::Bool
+    signatureHelpProvider::SignatureHelpOptions
+    documentSymbolProvider::Bool
+    # referencesProvider::Bool
+    # documentHighlightProvider::Bool
+    # workspaceSymbolProvider::Bool
+    # codeActionProvider::Bool
+    # codeLensProvider::CodeLensOptions
+    # documentFormattingProvider::Bool
+    # documentRangeFormattingProvider::Bool
+    # documentOnTypeFormattingProvider::DocumentOnTypeFormattingOptions
+    # renameProvider::Bool
+end
+
+const serverCapabilities = ServerCapabilities(
+                        TextDocumentSyncKind["Incremental"],
+                        true, #hoverProvider
+                        CompletionOptions(false,["."]),
+                        true, #definitionProvider
+                        SignatureHelpOptions(["("]),
+                        true) # documentSymbolProvider 
+
+type InitializeResult
+    capabilities::ServerCapabilities
+end
+
+function process(r::JSONRPC.Request{Val{Symbol("initialize")},Dict{String,Any}}, server)
+    server.rootPath=haskey(r.params,"rootPath") ? r.params["rootPath"] : ""
+    response = JSONRPC.Response(get(r.id), InitializeResult(serverCapabilities))
+    send(response, server)
+end
+
+function JSONRPC.parse_params(::Type{Val{Symbol("initialize")}}, params)
+    return Any(params)
+end
+
+
 type LanguageServerInstance
     pipe_in
     pipe_out
@@ -26,4 +78,13 @@ function Base.run(server::LanguageServerInstance)
 
         process(request, server)
     end
+end
+
+
+function process(r::JSONRPC.Request{Val{Symbol("\$/cancelRequest")},CancelParams}, server)
+    
+end
+
+function JSONRPC.parse_params(::Type{Val{Symbol("\$/cancelRequest")}}, params)
+    return CancelParams(params)
 end
