+++ src/provider_definitions.jl
@@ -1,3 +1,4 @@
+## provides function definition links ##
 function process(r::JSONRPC.Request{Val{Symbol("textDocument/definition")},TextDocumentPositionParams}, server)
     word = get_word(r.params, server)
     x = get_sym(word)
